moved from query handler to scheduled job
